Update versions and remove `sqlmodel`
Update almost all package versions. In doing so, I've encountered an
error caused by `SQLModel` not collecting metadata for public and
foreign keys. Since I found out that there is a more streamlined API in
SQLAlchemy called "declarative mapping," we are going to replace
`sqlmodel` with the that instead. Other changes include:
- data-star v1
- SSE responses using datastar-py
- CSS animation of modal elements without data-star hacks
- Fix for `doit run watch` failure previously caused by discarding the
  current environment variables
- Change the underlying ASGI server from Uvicorn to Granian with HTTP/2
  enabled